Extract Story out into its own module.
