+++ examples/minimal/components/heading/stories.py
@@ -1,8 +1,8 @@
 """The subject for this Heading component."""
 from pathlib import Path
 
-from storytime import Story
 from storytime import Subject
+from storytime.story import Story
 
 
 def this_subject() -> Subject:
