+++ src/storytime/__init__.py
@@ -17,9 +17,11 @@ from pathlib import Path
 from types import ModuleType
 from typing import get_type_hints
 from typing import Optional
+from typing import TYPE_CHECKING
 from typing import Union
 
-from viewdom.render import VDOM
+if TYPE_CHECKING:
+    from .story import Story
 
 
 def make_target_path(target_module: str) -> Path:
@@ -144,11 +146,3 @@ class Subject:
     title: str
     subject_path: Path
     stories: list[Story] = field(default_factory=list, hash=False)
-
-
-@dataclass(frozen=True)
-class Story:
-    """The actual contents of an actual story."""
-
-    title: str
-    template: Optional[VDOM] = None
